feature[test]: refactoring issue #352
