+++ mockk/common/src/test/kotlin/io/mockk/it/CapturingTest.kt
@@ -1,21 +1,19 @@
 package io.mockk.it
 
 import io.mockk.*
+import kotlin.test.BeforeTest
 import kotlin.test.Test
 import kotlin.test.assertEquals
+import kotlin.test.assertFailsWith
 
 class CapturingTest {
 
-    class Cls {
-        var value = 0
-    }
+    private val mock = mockk<MockedSubject>()
 
-    class MockCls {
-        fun op(a: Int, b: Int, c: Cls) = a + b + c.value
-    }
-
-    class CoMockCls {
-        suspend fun op(a: Int, b: Int, c: Cls) = a + b + c.value
+    @BeforeTest
+    fun setup() {
+        every { mock.doSomething("1", "data1") } returns "result1"
+        every { mock.doSomething("2", "data2") } returns "result2"
     }
 
     @Test
@@ -72,5 +70,91 @@ class CapturingTest {
         assertEquals(55, items[0].value)
 
         coVerify { mock.op(1, 2, any()) }
+    }
+
+    /**
+     * See issue #352.
+     */
+    @Test
+    fun itThrowsAMockkExceptionWhenVerifyingTheSameFunctionTwiceWithSlots() {
+        mock.doSomething("1", "data1")
+        mock.doSomething("2", "data2")
+
+        val dataSlotId1 = slot<String>()
+        val dataSlotId2 = slot<String>()
+
+        assertFailsWith<MockKException> {
+            verify {
+                mock.doSomething("1", capture(dataSlotId1))
+                mock.doSomething("2", capture(dataSlotId2))
+            }
+        }
+    }
+
+    /**
+     * See issue #352.
+     */
+    @Test
+    fun itDoesNotThrowAMockkExceptionWhenThereAreMultipleTestsVerifyingWithSlots() {
+        mock.doSomething("1", "data1")
+
+        val slot = slot<String>()
+        verify {
+            mock.doSomething("1", capture(slot))
+        }
+
+        assertEquals("data1", slot.captured)
+    }
+
+    /**
+     * See issue #352.
+     */
+    @Test
+    fun anotherTestToTestTheCoexistenceOfTestsWithSlots() {
+        mock.doSomething("1", "data1")
+
+        val slot = slot<String>()
+        verify {
+            mock.doSomething("1", capture(slot))
+        }
+
+        assertEquals("data1", slot.captured)
+    }
+
+    /**
+     * See issue #352.
+     */
+    @Test
+    fun itAllowsMultipleCapturingsOfTheSameFunctionUsingAMutableList() {
+        mock.doSomething("1", "data1")
+        mock.doSomething("2", "data2")
+
+        val slotList = mutableListOf<String>()
+
+        verify {
+            mock.doSomething("1", capture(slotList))
+            mock.doSomething("2", capture(slotList))
+        }
+
+        assertEquals("data1", slotList[0])
+        assertEquals("data2", slotList[1])
+    }
+
+    class Cls {
+        var value = 0
+    }
+
+    class MockCls {
+        fun op(a: Int, b: Int, c: Cls) = a + b + c.value
+    }
+
+    class CoMockCls {
+        suspend fun op(a: Int, b: Int, c: Cls) = a + b + c.value
+    }
+
+    open class MockedSubject {
+        open fun doSomething(id: String?, data: Any?): String {
+            throw IllegalStateException("Not mocked :(")
+        }
     }
 }
